requirements.txt
numpy
matplotlib
Pillow
segyio
tqdm
